frame: Re-jig structure for handling AX.25 2.2
AX.25 2.2 adds support for modulo-128 frames using a 16-bit control
field.  The only way to know you're dealing with modulo-128 traffic is
by remembering that you negotiated it earlier.

Thus we need 8-bit and 16-bit versions of the frame structure to handle
the enlarged control field in the modulo-128 case.